Replace SinkStorageReader and SinkStorageWriter with VirtualDocument[Tuple] (#3024)
This PR removes the `SinkStorageReader` and `SinkStorageWriter`, and
substitute it with `VirtualDocument[Tuple]`

### 1. Recap on `SinkStorage` and `VirtualDocument`

`SinkStorage` is one kinds of storage specialized for Tuple storage. It
relies on the `setSchema` function to know how to serialize/deserialize
the tuples, and has `getSchema` function to return out the stored
schema. In other words, `SinkStorage` stores both the tuples AND schema.

`VirtualDocument` is an abstraction layer for all the storage-related
apis. It currently has `MongoDocument[T]`. `MemoryDocument[T]`,
`FileDocument[T]` implemented. These implementations are designed to
store the actual data(a sequence of `T` objects), and `T`'s
serializer/deserializer if any. Compared to storing tuples like
`SinkStorage`, implementations will ONLY store tuples, NO schema,
because schema is very specialized for `TupleLike` object and should not
be part of the Document.

### 2. Changes in the `OpResultStorage`

`OpResultStorage` manages all the opened `SinkStorageReader`. So to
persist this part, the `OpResultStorage` extends from
`ConcurrentHashMap[OperatorIdentity, SinkStorageReader]` to
`ConcurrentHashMap[OperatorIdentity, (VirtualDocument[Tuple],
Option[Schema])]`. This extension is due to the fact that
`VirtualDocument[T]` doesn't have interfaces to acquire the `Schema`.
`Option` is used because in some cases the schema is not generated yet
when the `VirtualDocument` is initialized

### 3. New method `setSerde` in the `MongoDocument`

A new function is added:

```
  /**
    * set the serializer & deserializer.
    *
    * @param toDocument: the serializer, convert T to MongoDB's Document.
    * @param fromDocument: the deserializer, convert MongoDB's Document to T.
    */
  def setSerde(toDocument: T => Document, fromDocument: Document => T): Unit = {
    this.toDocument = Some(toDocument)
    this.fromDocument = Some(fromDocument)
  }
```

this is used to reset the document's the serializer/deserializer . This
is needed because in some cases, the document needs to be created, but
the serializer and deserializer are not ready yet.

### 4. New trait, `TupleResultStatistics`, is added

Since when MongoDB is used for result storage, the statistics of columns
will be shown in the result panel. A new trait is added:

```
package edu.uci.ics.amber.engine.common.storage

trait TupleResultStatistics {
  // a map from numerical fields to stats
  def getNumericColStats: Map[String, Map[String, Any]]

  // a map from date fields to stats
  def getDateColStats: Map[String, Map[String, Any]]

  // a map from categorical fields to stats
  def getCategoricalStats: Map[String, Map[String, Any]]
}
```

And MongoDocument implements this trait:
```
  override def getNumericColStats: Map[String, Map[String, Any]] =
    collectionMgr.calculateNumericStats()

  override def getDateColStats: Map[String, Map[String, Any]] = collectionMgr.calculateDateStats()

  override def getCategoricalStats: Map[String, Map[String, Any]] =
    collectionMgr.calculateCategoricalStats()
```